Implemented Database
Added the .edmx to the TaskDAL in order to persist data and gather the data from the SQL database.
Changed TaskList Page to gather data from SQL database rather than statically.
Added a SQL query to create database and table
*NOTE*
The SQL query screwed up and didn't make the id the identity.